:truck: moved functions that mutates the mail context from 	the `mail_context` module to the `action` module.
