+++ src/rules/modules/mail_context.rs
@@ -22,9 +22,8 @@ pub mod mail_context {
 
     use crate::{
         rules::address::Address, rules::modules::types::Rcpt, rules::modules::EngineResult,
-        rules::obj::Object, rules::server_api::ServerAPI, smtp::mail::Body,
+        rules::server_api::ServerAPI,
     };
-    use std::io::Write;
 
     // FIXME: all those poison error map_err make the code harder to read.
 
@@ -74,36 +73,6 @@ pub mod mail_context {
             .clone())
     }
 
-    #[rhai_fn(global, return_raw)]
-    pub fn rewrite_mail_from(
-        this: &mut std::sync::Arc<std::sync::RwLock<ServerAPI>>,
-        addr: String,
-    ) -> EngineResult<()> {
-        let addr = Address::new(&addr).map_err::<Box<EvalAltResult>, _>(|_| {
-            format!(
-                "could not rewrite mail_from with '{}' because it is not valid address",
-                addr,
-            )
-            .into()
-        })?;
-
-        let mut email = &mut this
-            .write()
-            .map_err::<Box<EvalAltResult>, _>(|e| e.to_string().into())?
-            .mail_context;
-
-        email.envelop.mail_from = addr.clone();
-
-        match &mut email.body {
-            Body::Empty => Err("failed to rewrite mail_from: the email has not been received yet. Use this method in postq or later.".into()),
-            Body::Raw(_) => Err("failed to rewrite mail_from: the email has not been parsed yet. Use this method in postq or later.".into()),
-            Body::Parsed(body) => {
-                body.rewrite_mail_from(addr.full());
-                Ok(())
-            },
-        }
-    }
-
     #[rhai_fn(global, get = "rcpt", return_raw)]
     pub fn rcpt(this: &mut std::sync::Arc<std::sync::RwLock<ServerAPI>>) -> EngineResult<Rcpt> {
         Ok(this
@@ -113,96 +82,6 @@ pub mod mail_context {
             .envelop
             .rcpt
             .clone())
-    }
-
-    #[rhai_fn(global, return_raw)]
-    pub fn rewrite_rcpt(
-        this: &mut std::sync::Arc<std::sync::RwLock<ServerAPI>>,
-        index: String,
-        addr: String,
-    ) -> EngineResult<()> {
-        let index = Address::new(&index).map_err::<Box<EvalAltResult>, _>(|_| {
-            format!(
-                "could not rewrite address '{}' because it is not valid address",
-                index,
-            )
-            .into()
-        })?;
-
-        let addr = Address::new(&addr).map_err::<Box<EvalAltResult>, _>(|_| {
-            format!(
-                "could not rewrite address '{}' with '{}' because it is not valid address",
-                index, addr,
-            )
-            .into()
-        })?;
-
-        let email = &mut this
-            .write()
-            .map_err::<Box<EvalAltResult>, _>(|e| e.to_string().into())?
-            .mail_context;
-
-        email.envelop.rcpt.remove(&index);
-        email.envelop.rcpt.insert(addr.clone());
-
-        match &mut email.body {
-            Body::Empty => Err("failed to rewrite rcpt: the email has not been received yet. Use this method in postq or later.".into()),
-            Body::Raw(_) => Err("failed to rewrite rcpt: the email has not been parsed yet. Use this method in postq or later.".into()),
-            Body::Parsed(body) => {
-                body.rewrite_rcpt(index.full(), addr.full());
-                Ok(())
-            },
-        }
-    }
-
-    #[rhai_fn(global, return_raw)]
-    pub fn add_rcpt(
-        this: &mut std::sync::Arc<std::sync::RwLock<ServerAPI>>,
-        s: String,
-    ) -> EngineResult<()> {
-        let new_addr = Address::new(&s)
-            .map_err(|_| format!("{} could not be converted to a valid rcpt address", s))?;
-
-        let email = &mut this
-            .write()
-            .map_err::<Box<EvalAltResult>, _>(|e| e.to_string().into())?
-            .mail_context;
-
-        email.envelop.rcpt.insert(new_addr.clone());
-
-        match &mut email.body {
-            Body::Empty => Err("failed to rewrite rcpt: the email has not been received yet. Use this method in postq or later.".into()),
-            Body::Raw(_) => Err("failed to rewrite rcpt: the email has not been parsed yet. Use this method in postq or later.".into()),
-            Body::Parsed(body) => {
-                body.add_rcpt(new_addr.full());
-                Ok(())
-            },
-        }
-    }
-
-    #[rhai_fn(global, return_raw)]
-    pub fn remove_rcpt(
-        this: std::sync::Arc<std::sync::RwLock<ServerAPI>>,
-        s: String,
-    ) -> EngineResult<()> {
-        let addr = Address::new(&s)
-            .map_err(|_| format!("{} could not be converted to a valid rcpt address", s))?;
-
-        let email = &mut this
-            .write()
-            .map_err::<Box<EvalAltResult>, _>(|e| e.to_string().into())?
-            .mail_context;
-
-        email.envelop.rcpt.remove(&addr);
-
-        match &mut email.body {
-            Body::Empty => Err("failed to rewrite rcpt: the email has not been received yet. Use this method in postq or later.".into()),
-            Body::Raw(_) => Err("failed to rewrite rcpt: the email has not been parsed yet. Use this method in postq or later.".into()),
-            Body::Parsed(body) => {
-                body.remove_rcpt(addr.full());
-                Ok(())
-            },
-        }
     }
 
     #[rhai_fn(global, get = "mail_timestamp", return_raw)]
@@ -273,254 +152,5 @@ pub mod mail_context {
             this.read()
                 .map_err::<Box<EvalAltResult>, _>(|e| e.to_string().into())?
         ))
-    }
-
-    /// write the current email to a specified folder.
-    #[rhai_fn(global, return_raw)]
-    pub fn write(
-        this: &mut std::sync::Arc<std::sync::RwLock<ServerAPI>>,
-        dir: &str,
-    ) -> EngineResult<()> {
-        match std::fs::OpenOptions::new().create(true).write(true).open(
-            std::path::PathBuf::from_iter([dir, &format!("{}.eml", message_id(this)?)]),
-        ) {
-            Ok(file) => {
-                let mut writer = std::io::LineWriter::new(file);
-
-                match &this
-                    .read()
-                    .map_err::<Box<EvalAltResult>, _>(|e| e.to_string().into())?
-                    .mail_context
-                    .body
-                {
-                    Body::Empty => {
-                        return Err(
-                            "failed to write email: the body has not been received yet.".into()
-                        )
-                    }
-                    Body::Raw(raw) => writer.write_all(raw.as_bytes()),
-                    Body::Parsed(email) => {
-                        let (headers, body) = email.to_raw();
-                        writer.write_all(format!("{}\n\n{}", headers, body).as_bytes())
-                    }
-                }
-            }
-            .map_err(|err| format!("failed to write email at '{}': {:?}", dir, err).into()),
-            Err(err) => Err(format!("failed to write email at '{}': {:?}", dir, err).into()),
-        }
-    }
-
-    /// write the content of the current email in a json file.
-    /// NOTE: it would be great not having all those 'map_err'.
-    #[rhai_fn(global, return_raw)]
-    pub fn dump(
-        this: &mut std::sync::Arc<std::sync::RwLock<ServerAPI>>,
-        dir: &str,
-    ) -> EngineResult<()> {
-        match std::fs::OpenOptions::new().create(true).write(true).open(
-            std::path::PathBuf::from_iter([dir, &format!("{}.dump.json", message_id(this)?)]),
-        ) {
-            Ok(mut file) => file
-                .write_all(
-                    serde_json::to_string_pretty(
-                        &this
-                            .read()
-                            .map_err::<Box<EvalAltResult>, _>(|e| e.to_string().into())?
-                            .mail_context,
-                    )
-                    .map_err::<Box<EvalAltResult>, _>(|err| {
-                        format!("failed to dump email at '{}': {:?}", dir, err).into()
-                    })?
-                    .as_bytes(),
-                )
-                .map_err(|err| format!("failed to dump email at '{}': {:?}", dir, err).into()),
-            Err(err) => Err(format!("failed to dump email at '{}': {:?}", dir, err).into()),
-        }
-    }
-
-    // TODO: unfinished, queue parameter should point to a folder specified in toml config.
-    /// dump the current email into a quarantine queue, skipping delivery.
-    #[rhai_fn(global, return_raw)]
-    pub fn quarantine(
-        this: &mut std::sync::Arc<std::sync::RwLock<ServerAPI>>,
-        queue: &str,
-    ) -> EngineResult<()> {
-        match std::fs::OpenOptions::new()
-            .create(true)
-            .append(true)
-            .open(queue)
-        {
-            Ok(mut file) => {
-                disable_delivery(this)?;
-
-                file.write_all(
-                    serde_json::to_string_pretty(
-                        &this
-                            .write()
-                            .map_err::<Box<EvalAltResult>, _>(|err| {
-                                format!("failed to dump email: {err:?}").into()
-                            })?
-                            .mail_context,
-                    )
-                    .map_err::<Box<EvalAltResult>, _>(|err| {
-                        format!("failed to dump email: {err:?}").into()
-                    })?
-                    .as_bytes(),
-                )
-                .map_err(|err| format!("failed to dump email: {err:?}").into())
-            }
-            Err(err) => Err(format!("failed to dump email: {err:?}").into()),
-        }
-    }
-
-    #[rhai_fn(global, return_raw)]
-    pub fn deliver(
-        this: &mut std::sync::Arc<std::sync::RwLock<ServerAPI>>,
-        resolver: String,
-    ) -> EngineResult<()> {
-        this.write()
-            .map_err::<Box<EvalAltResult>, _>(|e| e.to_string().into())?
-            .mail_context
-            .metadata
-            .as_mut()
-            .ok_or_else::<Box<EvalAltResult>, _>(|| {
-                "failed to use default resolver: email metadata are unavailable".into()
-            })?
-            .resolver = resolver;
-
-        Ok(())
-    }
-
-    #[rhai_fn(global, return_raw)]
-    pub fn disable_delivery(
-        this: &mut std::sync::Arc<std::sync::RwLock<ServerAPI>>,
-    ) -> EngineResult<()> {
-        deliver(this, "none".to_string())
-    }
-
-    /// check if a given header exists in the top level headers.
-    #[rhai_fn(global, return_raw, pure)]
-    pub fn has_header(
-        this: &mut std::sync::Arc<std::sync::RwLock<ServerAPI>>,
-        header: &str,
-    ) -> EngineResult<bool> {
-        Ok(
-            match &this
-                .read()
-                .map_err::<Box<EvalAltResult>, _>(|e| e.to_string().into())?
-                .mail_context
-                .body
-            {
-                Body::Empty => false,
-                Body::Raw(raw) => {
-                    let mut headers_end = 0;
-
-                    // getting headers from the raw email.
-                    for line in raw.lines() {
-                        let mut split = line.splitn(2, ':');
-                        match (split.next(), split.next()) {
-                            // adding one to the index because `\n` is striped using the Lines iterator.
-                            (Some(_), Some(_)) => headers_end += line.len() + 1,
-                            _ => break,
-                        }
-                    }
-
-                    raw[0..headers_end].contains(format!("{}: ", header).as_str())
-                }
-                Body::Parsed(email) => email.headers.iter().any(|(name, _)| header == name),
-            },
-        )
-    }
-
-    /// add a header to the raw or parsed email contained in ctx.
-    #[rhai_fn(global, return_raw)]
-    pub fn add_header(
-        this: &mut std::sync::Arc<std::sync::RwLock<ServerAPI>>,
-        header: &str,
-        value: &str,
-    ) -> EngineResult<()> {
-        match &mut this
-            .write()
-            .map_err::<Box<EvalAltResult>, _>(|e| e.to_string().into())?
-            .mail_context
-            .body
-        {
-            Body::Empty => {
-                return Err(format!(
-                    "failed to add header '{}': the body has not been received yet.",
-                    header
-                )
-                .into())
-            }
-            Body::Raw(raw) => *raw = format!("{}: {}\n{}", header, value, raw),
-            Body::Parsed(email) => email.headers.push((header.to_string(), value.to_string())),
-        };
-
-        Ok(())
-    }
-
-    /// add a recipient to the list recipient using a raw string.
-    #[rhai_fn(global, name = "bcc", return_raw)]
-    pub fn bcc_str(
-        this: &mut std::sync::Arc<std::sync::RwLock<ServerAPI>>,
-        bcc: &str,
-    ) -> EngineResult<()> {
-        this.write()
-            .map_err::<Box<EvalAltResult>, _>(|e| e.to_string().into())?
-            .mail_context
-            .envelop
-            .rcpt
-            .insert(Address::new(bcc).map_err(|_| {
-                format!("'{}' could not be converted to a valid rcpt address", bcc)
-            })?);
-
-        Ok(())
-    }
-
-    /// add a recipient to the list recipient using an address.
-    #[rhai_fn(global, name = "bcc", return_raw)]
-    pub fn bcc_addr(
-        this: &mut std::sync::Arc<std::sync::RwLock<ServerAPI>>,
-        bcc: Address,
-    ) -> EngineResult<()> {
-        this.write()
-            .map_err::<Box<EvalAltResult>, _>(|e| e.to_string().into())?
-            .mail_context
-            .envelop
-            .rcpt
-            .insert(bcc);
-
-        Ok(())
-    }
-
-    /// add a recipient to the list recipient using an object.
-    #[rhai_fn(global, name = "bcc", return_raw)]
-    pub fn bcc_object(
-        this: &mut std::sync::Arc<std::sync::RwLock<ServerAPI>>,
-        bcc: std::sync::Arc<Object>,
-    ) -> EngineResult<()> {
-        this.write()
-            .map_err::<Box<EvalAltResult>, _>(|e| e.to_string().into())?
-            .mail_context
-            .envelop
-            .rcpt
-            .insert(match &*bcc {
-                Object::Address(addr) => addr.clone(),
-                Object::Str(string) => Address::new(string.as_str()).map_err(|_| {
-                    format!(
-                        "'{}' could not be converted to a valid rcpt address",
-                        string
-                    )
-                })?,
-                other => {
-                    return Err(format!(
-                        "'{}' could not be converted to a valid rcpt address",
-                        other.to_string()
-                    )
-                    .into())
-                }
-            });
-
-        Ok(())
     }
 }
